requirements.txt
alabaster==0.7.3
altgraph==0.10.2
appnope==0.1.0
astroid==1.3.6
Babel==1.3
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
beautifulsoup4==4.3.2
bonjour-py==0.3
chardet==2.3.0
click==6.6
coverage==4.0a5
coveralls==1.1
cssselect==0.9.2
curses-menu==0.5.0
decorator==4.0.10
Django==1.7.6
django-bootstrap-form==3.2
django-bootstrap-forms==0.1
django-bower==5.0.2
django-cors-headers==1.0.0
django-nose==1.3
djangorestframework==3.1.1
docopt==0.6.2
docutils==0.12
Flask==0.11.1
Flask-Cors==2.1.2
Flask-GoogleMaps==0.2.0
FormEncode==1.3.0
future==0.15.2
geopy==1.11.0
googlemaps==2.2
gpsoauth==0.3.0
httplib2==0.9
httpretty==0.8.8
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
logilab-common==0.63.2
lxml==3.6.0
macholib==1.5.1
mailchimp3==1.0.17
MarkupSafe==0.23
matplotlib==1.3.1
mock==1.0.1
modulegraph==0.10.4
mysql==0.0.1
MySQL-python==1.2.5
nose==1.3.4
numpy==1.8.0rc1
oauth2==1.5.211
oauthlib==0.7.2
pathlib2==2.1.0
peewee==2.8.1
pexpect==4.2.0
pickleshare==0.7.2
prompt-toolkit==1.0.3
protobuf==2.6.1
protobuf-to-dict==0.1.0
psutil==4.3.0
ptyprocess==0.5.1
py2app==0.7.3
pycryptodomex==3.4.2
PyDispatcher==2.0.5
Pygments==2.0.2
pylint==1.4.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
PyPrind==2.9.8
python-dateutil==1.5
python-openid==2.2.5
PyTumblr==0.0.6
pytz==2013.7
requests==2.10.0
requests-oauthlib==0.4.2
s2sphere==0.2.4
scipy==0.13.0b1
simplegeneric==0.8.1
simplejson==3.6.5
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
SQLObject==3.0.0
tensorflow==0.9.0
tqdm==4.8.3
traitlets==4.2.2
twilio==3.8.0
virtualenv==15.0.2
wcwidth==0.1.7
Werkzeug==0.11.10
xattr==0.6.4
zope.interface==4.1.1
